+++ CHANGELOG.md
@@ -1,2 +1,8 @@
+# 2.0.0
+
+- Now building for 1.20.1 (DC v6)
+- Fix Engineered Schematics main hand rendering
+- Fix Physics Mod EFT Rendering 
+
 # 1.0.8
 - Support older versions of LostCities